BACK-29 feat: add tank allocation endpoint
