+++ forecast/router.py
@@ -5,6 +5,7 @@ from sqlalchemy.orm import Session
 from database.database import get_db
 from delivery import crud
 from forecast.winters.RunWinters import RunWinters
+from forecast.tank_allocation import process_tank_data
 
 router = APIRouter(
     prefix="/forecasts",
@@ -18,3 +19,27 @@ def calculate_forecasts_for_every_fuel(station_id: int, db: Session = Depends(ge
     data = crud.read_deliveries_for_winters(station_id=station_id, limit=350, db=db)
     run_winters = RunWinters()
     return run_winters.for_every_fuel(data)
+
+
+@router.get("/tank_allocation/{station_id}")
+def calculate_tank_allocation(station_id: int, db: Session = Depends(get_db)):
+    data = {
+        "ulg95": {
+            "1": 1000, "2": 1000, "3": 1000, "4": 1000, "5": 1000,
+            "6": 1000, "7": 1000, "8": 1000, "9": 1000, "10": 1000
+        },
+        "dk": {
+            "1": 1000, "2": 1000, "3": 1000, "4": 1000, "5": 1000,
+            "6": 1000, "7": 1000, "8": 1000, "9": 1000, "10": 1000
+        },
+        "ultsu": {
+            "1": 1000, "2": 1000, "3": 1000, "4": 1000, "5": 1000,
+            "6": 1000, "7": 1000, "8": 1000, "9": 1000, "10": 1000
+        },
+        "ultdk": {
+            "1": 1000, "2": 1000, "3": 1000, "4": 1000, "5": 1000,
+            "6": 1000, "7": 1000, "8": 1000, "9": 1000, "10": 1000
+        }
+    }
+    start_date_str = "2024-01-01"
+    return process_tank_data(data, start_date_str, station_id)
